+++ src/test/java/oshi/SystemInfoTest.java
@@ -1,17 +1,18 @@
 /**
  * Oshi (https://github.com/dblock/oshi)
- * 
+ *
  * Copyright (c) 2010 - 2016 The Oshi Project Team
- * 
+ *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
- * 
+ *
  * Contributors:
  * dblock[at]dblock[dot]org
  * alessandro[at]perucchi[dot]org
  * widdis[at]gmail[dot]com
+ * enrico[dot]bianchi[at]gmail[dot]com
  * https://github.com/dblock/oshi/graphs/contributors
  */
 package oshi;
@@ -83,9 +84,16 @@ public class SystemInfoTest {
         HardwareAbstractionLayer hal = si.getHardware();
         GlobalMemory memory = hal.getMemory();
         assertNotNull(memory);
+        
+        // RAM tests
         assertTrue(memory.getTotal() > 0);
         assertTrue(memory.getAvailable() >= 0);
         assertTrue(memory.getAvailable() <= memory.getTotal());
+        
+        // Swap tests
+        assertTrue(memory.getSwapTotal() >= 0);
+        assertTrue(memory.getSwapUsed() >= 0);
+        assertTrue(memory.getSwapUsed() <= memory.getSwapTotal());
     }
 
     /**