+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 2.4 (in-progress)
 ================
 * [#140](https://github.com/dblock/oshi/pull/140): Added process and thread counts - [@dbwiddis](https://github.com/dbwiddis).
+* [#142](https://github.com/dblock/oshi/pull/142): Added methods for getting swap (total and used) in the system - [@henryx](https://github.com/henryx).
+
 * Your contribution here.
 
 2.3 (4/14/2016)